refactor: Inputer 공통화
Car, Moving Inputer가 PositiveNumInputer를 상속받아 공통화함
