+++ src/main/java/edu/nextstep/camp/racing/inputer/CarInputer.java
@@ -2,30 +2,18 @@ package edu.nextstep.camp.racing.inputer;
 
 import edu.nextstep.camp.racing.io.InputView;
 
-public class CarInputer {
+public class CarInputer extends PositiveNumInputer {
 
     private static final int MAX_CAR = 20;
     private static final int MIN_CAR = 2;
 
-    public InputPositiveResult input() {
-        try {
-            InputPositiveResult positiveResult = new InputPositiveResult(InputView.input());
-            checkInRangeByResultNum(positiveResult.result());
-            InputView.close();
-            return positiveResult;
-        } catch (RuntimeException e) {
-            System.out.println(e.getMessage());
-            return input();
-        }
+    @Override
+    protected int maxNum() {
+        return MAX_CAR;
     }
 
-    private void checkInRangeByResultNum(int resultNum) {
-        if (!isInRange(resultNum)) {
-            throw new IllegalArgumentException("Please input a number between "+MIN_CAR+" and "+MAX_CAR);
-        }
-    }
-
-    private boolean isInRange(int num) {
-        return MAX_CAR >= num && num >= MIN_CAR;
+    @Override
+    protected int minNum() {
+        return MIN_CAR;
     }
 }
